Oreobot design 2.0
Due to several adversities, we decided that it would be an intelligent decision to have a backup robot in case ours fails or has any problem during the competition, so we decided to use an EV3 Lego robot that was in our school and made another robot in case of emergence. This robot is our new son, Oreobot 2.0, and here is a brief description of it.